[FE-219] feat: 메인페이지 카테고리ID 유지 추가
